Introduce getEffectiveUsers, and improve the logic of checking usages
